fix(length-controls): border-radius accepts any CSS unit, not just px
W3-d rows 1-2 of 9 (contract §4.3). sgs/label.borderRadius and
sgs/product-card.tagBorderRadius now use a UnitControl and store a CSS length.

THREE-LAYER migration, not a control swap. C5's triage recorded these as
"simple JSX/prop swaps"; at source they were `type: number` attrs with numeric
render paths, so swapping the control alone stores a string into a number-typed
attr - which WP silently discards (D338).

- block.json x2: type -> [ "string", "number" ] (see LIVE BUG below).
- helpers-box.php: the SHARED sgs_label_box_css_rule() emitter replaced
  `intval($radius).'px'` with sgs_css_length_sanitise(). intval('1.5rem') is 1,
  so every non-px unit was silently truncated to 1px. One emitter, both callers
  (label/render.php:273, product-card/render.php:373) - universal per R-31-9.
- render.php x2: intval -> floatval on the zero-is-absent guard.
  intval('0.5rem') is 0, so every sub-1 rem/em radius was dropped entirely.
- edit.js x2: RangeControl -> UnitControl; unused imports removed.
- src/utils/cssLength.js: ONE definition of bare-number-means-px and
  zero-means-absent, shared by the canvas preview, mirrored by the PHP.

⛔ A REAL BUG THE LIVE CAPTURE CAUGHT, and the reason this took two deploys.
The first revision declared these attrs `"type": "string"`. Deployed and probed,
the legacy int 16 stored correctly in post_content but rendered NO radius: WP
validates against the declared type and falls back to the default on a mismatch.
EVERY EXISTING LABEL WITH A RADIUS WOULD HAVE SILENTLY LOST IT. The back-compat
claim in my earlier draft was false. The build was green, all 43 prebuild gates
passed, and the editor looked right - only a live capture falsified it. Fixed
with the union type, re-deployed, re-measured.

VERIFIED LIVE (page 2363, lifted stylesheet - the CSS is lifted to
uploads/sgs-css/, so a page-HTML grep proves nothing):
  legacy 16  -> border-radius:16px     (was: 16px, unchanged)
  "1.5rem"   -> border-radius:1.5rem   (was: 1px, truncated)
  "0.5rem"   -> border-radius:0.5rem   (was: dropped entirely)
  0          -> no declaration          (unchanged)
PHP diagnostics: 0.

Also fixes an editor-canvas bug found on the way: the old preview appended 'px'
unconditionally, so a rem value painted as '1.5rempx' and did nothing.

Reports: reports/visual-diff/{label,product-card}-2026-08-13.md. The
product-card report states plainly that its evidence is shared-emitter
equivalence with the label capture, not a second independent measurement.

+++ plugins/sgs-blocks/src/blocks/label/render.php
@@ -243,7 +243,13 @@ $base_padding_shorthand = sgs_box_object_shorthand( $padding_obj );
 // A meaningful (non-zero, present) border-radius. A stored 0 is treated as
 // "no rounding" and not emitted — keeps a bare eyebrow free of a pointless
 // `border-radius:0px` (regression guard: bare eyebrows stay box-free).
-$has_radius   = ( '' !== (string) $border_radius && 0 !== intval( $border_radius ) );
+//
+// floatval, NOT intval (2026-08-13): borderRadius is now a CSS-length STRING,
+// and intval('0.5rem') is 0 — which would silently DROP every sub-1 rem/em
+// radius. floatval('0.5rem') is 0.5, while floatval('0px') is still 0.0, so the
+// zero-is-absent guard above keeps working for both the legacy bare number and
+// an explicit zero length.
+$has_radius   = ( '' !== (string) $border_radius && 0.0 !== floatval( $border_radius ) );
 $radius_value = $has_radius ? $border_radius : '';
 
 // Box-present = either background channel (native style.color.background OR the
